# clinical-mdr-api/consumer_api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "StudyBuilder Consumer API",
...
    "version": "0.1.90"
  },
  "paths": {
...
    "/v1/studies": {
      "get": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Get Studies",
        "description": "Returns a paginated list of studies, sorted by the specified sort criteria and order.\n\nEach returned study contains a full list of corresponding study versions, sorted by version start date in descending order.\n\nReturned `version_number` value can be used in other endpoints to retrieve study entities (e.g. visits, activities, etc.)\nassociated with a specific study version.",
        "operationId": "get_studies_v1_studies_get",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          },
          {
            "BearerJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortByStudies",
              "default": "uid"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page Number"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by study ID (case-insensitive partial match), for example `NN1234-5678`.",
              "title": "Id"
            },
            "description": "Filter by study ID (case-insensitive partial match), for example `NN1234-5678`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Study_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/studies/{uid}/study-visits": {
      "get": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Get Study Visits",
        "description": "Returns a paginated list of study visits, sorted by the specified sort criteria and order.\n\nIf `study_version_number` query parameter is provided, study visits\nassociated with the specified study version will be returned.\nOtherwise, visits for the latest study version will be returned.",
        "operationId": "get_study_visits_v1_studies__uid__study_visits_get",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          },
          {
            "BearerJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Study UID",
              "title": "Uid"
            },
            "description": "Study UID"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortByStudyVisits",
              "default": "unique_visit_number"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page Number"
            }
          },
          {
            "name": "study_version_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Study Version Number",
              "title": "Study Version Number"
            },
            "description": "Study Version Number",
            "example": "2.1"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseWithStudyVersion_StudyVisit_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/studies/{uid}/study-activities": {
      "get": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Get Study Activities",
        "description": "Returns a paginated list of study activities, sorted by the specified sort criteria and order.\n\nIf `study_version_number` query parameter is provided, study activities\nassociated with the specified study version will be returned.\nOtherwise, activities for the latest study version will be returned.",
        "operationId": "get_study_activities_v1_studies__uid__study_activities_get",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          },
          {
            "BearerJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Study UID",
              "title": "Uid"
            },
            "description": "Study UID"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortByStudyActivities",
              "default": "uid"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Page Size"
            }
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page Number"
            }
          },
          {
            "name": "study_version_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Study Version Number",
              "title": "Study Version Number"
            },
            "description": "Study Version Number",
            "example": "2.1"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseWithStudyVersion_StudyActivity_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Exception class."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "More information about the error."
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time",
            "description": "The point in time when the error occurred."
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "The url/path of the request."
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "The HTTP method of the request."
          }
        },
        "type": "object",
        "required": [
          "type",
          "message",
          "time",
          "path",
          "method"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponseWithStudyVersion_StudyVisit_": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self",
            "description": "Pagination link pointing to the current page"
          },
          "prev": {
            "type": "string",
            "title": "Prev",
            "description": "Pagination link pointing to the previous page"
          },
          "next": {
            "type": "string",
            "title": "Next",
            "description": "Pagination link pointing to the next page"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyVisit"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items"
          },
          "study_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudyVersionSimple"
              },
              {
                "type": "null"
              }
            ],
            "description": "Study version information"
          }
        },
        "type": "object",
        "required": [
          "self",
          "prev",
          "next",
          "items"
        ],
        "title": "PaginatedResponseWithStudyVersion[StudyVisit]"
      },
...
      "PaginatedResponse_Study_": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self",
            "description": "Pagination link pointing to the current page"
          },
          "prev": {
            "type": "string",
            "title": "Prev",
            "description": "Pagination link pointing to the previous page"
          },
          "next": {
            "type": "string",
            "title": "Next",
            "description": "Pagination link pointing to the next page"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Study"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items"
          }
        },
        "type": "object",
        "required": [
          "self",
          "prev",
          "next",
          "items"
        ],
        "title": "PaginatedResponse[Study]"
      },
...
      "SortByStudies": {
        "type": "string",
        "enum": [
          "uid",
          "id_prefix",
          "number"
        ],
        "title": "SortByStudies"
      },
      "SortByStudyActivities": {
        "type": "string",
        "enum": [
          "uid",
          "activity_name"
        ],
        "title": "SortByStudyActivities"
      },
...
      "SortByStudyVisits": {
        "type": "string",
        "enum": [
          "uid",
          "visit_name",
          "unique_visit_number"
        ],
        "title": "SortByStudyVisits"
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
...
    },
    "securitySchemes": {
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "api:///API.call": "Make calls to the API"
            },
            "authorizationUrl": "",
            "tokenUrl": ""
          }
        }
      },
      "BearerJwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "description": "Access token that will be sent as `Authorization: Bearer {token}` header in all requests"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}